feat(*): makefedit exit with 1 without MAKEF_PATh
This commit modifies the makef function to check if makefpath returns a value before running make with the -f flag. If makefpath does not return a value, then make is run without the -f flag. It also modifies the makefpath function to return ./Makefile if MAKEF_PATH is not set and the ./Makefile file exists. Finally, it modifies the makefedit function to print an error message and return an error code if MAKEF_PATH is not set.